requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1